service tool renaming
