+++ src/app/services/inputTool.service.ts
